# Optimize log4cxx log level access 20210408