--- a/src/libopenrave/libopenrave.cpp
+++ b/src/libopenrave/libopenrave.cpp
@@ -570,7 +570,8 @@ public:
             if( bRead ) {
                 if( !!ifstream(fullfilename.c_str()) ) {
                     return fullfilename;
-                }            }
+                }
+            }
             else {
                 return fullfilename;
             }