Revives getting log level access from log4cxx by speeding up previous performance when using log4cxx.

Performance test script:
```c++
#include <iostream>

#include <mujincontrollercommon/logging.h>
#include <openrave/openrave.h>

MUJIN_LOGGER("mujin.robotbridges.server.core");


int main() {
    const size_t numIterations = 100000000;
    struct timespec start, end;

    /*** Benchmark openrave get log level ***/
    // warm up
    OpenRAVE::RaveInitialize(true);
    for (size_t itr = 0; itr < numIterations; itr++)
    {
        OpenRAVE::RaveGetDebugLevel();
    }
    // benchmark
    clock_gettime(CLOCK_MONOTONIC_RAW, &start);
    for (size_t itr = 0; itr < numIterations; itr++)
    {
        OpenRAVE::RaveGetDebugLevel();
    }
    clock_gettime(CLOCK_MONOTONIC_RAW, &end);
    time_t elapsed = (end.tv_sec - start.tv_sec) * 1000000000 + (end.tv_nsec - start.tv_nsec);

    // logging
    std::cout << "iterations " << numIterations << std::endl;
    std::cout << "total " << elapsed << " nsec" << std::endl;
    std::cout << "average " << elapsed / numIterations << " nsec" << std::endl;
    return 0;
}
```

current get log access without log4cxx
```c++
iterations 100000000
total 514619209 nsec
average 5 nsec
```
with previous get log access from log4cxx: https://github.com/rdiankov/openrave/pull/941/files
```c++
iterations 100000000
total 5940225472 nsec
average 59 nsec
```

with this MR changes to get log access from log4cxx by checking effective level
```c++
iterations 100000000
total 783700733 nsec
average 7 nsec
```